plugin/file: guard against cname loops
This adds a loop counter to the context, which is checked. If we get
more than 8 loops, we break the off the lookup. In the case of *file*
we return a server failure to prevent sending an illegal answer on the
wire.

Document the context in plugin.md as well: Key and LoopKey.

Adding the following comment in the code:

~~~
Note the answer we're returning will be incomplete (more cnames to be followed) or
illegal (wildcard cname with multiple identical records). For now it's more important
to protect ourselves then to give the client a valid answer.

Note under the moto of garbage in, garbage out, we may actually _not_ care.
~~~

Signed-off-by: Miek Gieben <miek@miek.nl>